+++ src/HttpHeader.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: HttpHeader.cc,v 1.73 2001/10/17 21:37:28 hno Exp $
+ * $Id: HttpHeader.cc,v 1.74 2001/10/24 08:19:07 hno Exp $
  *
  * DEBUG: section 55    HTTP Header
  * AUTHOR: Alex Rousskov
@@ -331,7 +331,7 @@ httpHeaderClean(HttpHeader * hdr)
 	/* tmp hack to try to avoid coredumps */
 	if (e->id < 0 || e->id >= HDR_ENUM_END) {
 	    debug(55, 0) ("httpHeaderClean BUG: entry[%d] is invalid (%d). Ignored.\n",
-		pos, e->id);
+		(int) pos, e->id);
 	} else {
 	    statHistCount(&HttpHeaderStats[hdr->owner].fieldTypeDistr, e->id);
 	    /* yes, this destroy() leaves us in an incosistent state */
@@ -1014,7 +1014,7 @@ httpHeaderEntryParseCreate(const char *field_start, const char *field_end)
     if (field_end - value_start > 65536) {
 	/* String has a 64K limit */
 	debug(55, 1) ("WARNING: ignoring '%s' header of %d bytes\n",
-	    strBuf(e->name), field_end - value_start);
+	    strBuf(e->name), (int) (field_end - value_start));
 	if (e->id == HDR_OTHER)
 	    stringClean(&e->name);
 	memFree(e, MEM_HTTP_HDR_ENTRY);
